refactor: クライアント認証アクションを auth/actions.ts へ移す
`src/lib/auth.ts` (クライアント側の signInWithGoogle / signOut) と
`src/lib/auth/auth.ts` (サーバ側の betterAuth インスタンス) がほぼ同名で並んで
いた。片方は必ずクライアント、もう片方は必ずサーバという境界なので、誤って
import した場合のコストが大きい。

なぜ改名か: 名前で役割が判別できない状態は、レビューでも型でも捕まらない。
`auth/actions.ts` にして、同じディレクトリ内で auth-client / auth / session /
actions が役割ごとに並ぶ形にした。内容は移動のみで変更なし。
