+++ src/components/shared/header/user-menu/UserMenu.tsx
@@ -10,7 +10,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import type { UserWithEmail } from "@/entities/user";
-import { signOut } from "@/lib/auth";
+import { signOut } from "@/lib/auth/actions";
 
 // similarity-ignore: コンポーネント固有の Props 契約。構造が `{ user }` と偶然一致するが責務は別。
 type UserMenuProps = {
